[+] Fix bugs of empire reverse tcp exploit for windows
